ProbabilisticEchoInversion

Documentation for ProbabilisticEchoInversion.

ProbabilisticEchoInversion.MAPMCMCSolver

ProbabilisticEchoInversion.MAPSolver

ProbabilisticEchoInversion.MCMCSolver

ProbabilisticEchoInversion.apes

ProbabilisticEchoInversion.iterspectra

ProbabilisticEchoInversion.mapspectra

ProbabilisticEchoInversion.solve

ProbabilisticEchoInversion.MAPMCMCSolver — Type

MAPMCMCSolver([;optimizer, options])

Construct an MAPMCMCSolver, specifying how to invert a probabilistic backscattering model using a combination of maximum a posteriori optimization and Markov-chain Monte Carlo. This simply means that an optimization routine finds the MAP point estimate of the parameters, which is then used as the starting point for the MCMC run.

Arguments correspond exactly to the ones for MAPSolver and MCMCSolver; refer to their documentation for details.

ProbabilisticEchoInversion.MAPSolver — Type

MAPSolver([;optimizer, options])

Construct a MAPSolver, specifying how to invert a probabilistic backsattering model using maximum a-posteriori optimization. Default optimizer is L-BFGS. See Turing.jl and Optim.jl documentation for more information on available solvers and options.

ProbabilisticEchoInversion.MCMCSolver — Type

MCMCSolver([;sampler, parallel, nsamples, nchains; kwargs, verbose])

Construct an MCMCSolver, specifying how to invert a probabilistic backscattering model using Markov-chain Monte Carlo. By default uses the no-U-turn sampler with acceptance rate 0.8 and collects 1000 samples. See Turing.jl documentation for more information on options for MCMC sampling.

ProbabilisticEchoInversion.apes — Method

apes(echogram, model, solver[; params, result_handler, safe_precision, distributed])

Run the automatic probabilistic echo solver defined by the inverse model and solution method solver on the acoustic backstter data in echogram.

Arguments

echogram::DimArray: Acoustic backscatter data in the linear domain (i.e., volume backsattering coefficient $s_v$, area backsattering coefficient $s_a$, or nautical area scattering coefficient, $NASC$). The last dimension of the DimArray should be named :F and index the acoustic frequencies; all other dimensions should reference spatial/temporal coordinates.

model::Function: Probabilistic inverse model defined with Turing.jl or DynamicPPL.jl. This model should have the signature model(data, params), where data and params contain the acoustic data and any additional parameters. See below for more details.

solver::AbstractSolver: The method used to solve the inverse problem specified in model. See MCMCSolver and MAPSolver for more detail.

params: Optional additional params to pass to model.

result_handler: Optional function to transform the output of the solver before (for instance, by calculating the means of a Markov chain).

distributed::Bool=false: Whether to use all available processors when fitting model to echogram cells.

Details

This function applies a probabilistic inverse backscattering model defined using Turing.jl to each spectrum in a mulifrequency echogram. The model's constructor must accept two arguments:

data: A NamedTuple or other structure accessible by dot-notation, with fields coords, freqs, and backscatter. These contain the observed acoustic data. The model doesn't have to use all of them.

params: Optional NamedTuple or other object, containing any constants or auxiliary information used by the model.

ProbabilisticEchoInversion.iterspectra — Function

iterspectra(echogram[, freqdim])

Given an mulifrequency or broadband echogram in the form of a DimArray, with the acoustic frequencies in one dimension (by default named :F), iterate over each spectrum. The iterator yields NamedTuples with three fields:

coords: Coordinates of the spectrum on the non-:F dimensions of the Array (i.e. its location space/time)

freqs: Array of acoustic frequencies

backscatter: Array of backscatter values

ProbabilisticEchoInversion.mapspectra — Method

mapspectra(f, echogram[; freqdim, distributed])

Map the function f over each spectrum in the DimArray echogram. By default assumes the acoustic frequencies are recorded in dimension :F, if this is not the case, specify the name of the dimension using the freqdim argument.

ProbabilisticEchoInversion.solve — Function

solve(data, model, solver[, params])

Run the probabilistic inverse model defined by model on the acoustic backscatter spectrum in data, using solver as the inference engine.